Fixed an issue with functional ops and precidence
I added some more complex tests and found out that, because the math operators had a greater precidence than functions, that the AST would be parsed in such a way where the operands for a math operator could be a lambda that was meant for a function rather than the result of the function itself.

See the examples I checked in.