Fix member access precedence
- It seems that function calling and indexing needs to be implemented as
an operator so that it plays nicely with the other operators.
- Calling a function (or indexing) is techincally a binary operator
where the left side is the target and the right side is the arguments.
However, there is not usually a binary operator assigned to these
concepts. The way that I have got the parsing to work is to use
look-ahead to find the start token for a function call or index and then
pretend that it was a binary operator.
- The last piece of the puzzle was to make sure that function calling,
array subscripting and member access all had the same precedence.
- I should probably ask someone more knowledgeable than me if this is an
efficient or idiomatic way to do this.