Improved list pattern matching
- Pattern matching now has an explicit operator (^) to make it easier to
distinguish when parsing
- Pattern matching is now recursive so you can do things like mix
variable assignment and literal matching (e.g. ^[a, 2, c] = [1, 2, 3])

There are some improvements that need to be done before merging:

- Clean up the complicated pattern match functions (at least extract
some things into functions)
- Come up with a good syntax for map pattern matching (e.g. matching on
literals as well as variables)
- Test and fix issues with nested pattern matching (e.g. ^[^[^[a]]] =
[[[1]]])
- Work out whether the last variable in a match should be wrapped in an
array if the number of bindings are equal to the number of elements in
an array (e.g. ^[a, b, c] = [1, 2, 3] will currently assign [3] to c)